Moved classes out of internal package
