+++ geode-assembly/src/test/java/org/apache/geode/management/internal/cli/commands/GfshStartLocatorLogTest.java
@@ -25,9 +25,9 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
 import org.apache.geode.internal.AbstractConfig;
-import org.apache.geode.test.internal.rules.gfsh.GfshExecution;
-import org.apache.geode.test.internal.rules.gfsh.GfshRule;
-import org.apache.geode.test.internal.rules.gfsh.GfshScript;
+import org.apache.geode.test.rules.gfsh.GfshExecution;
+import org.apache.geode.test.rules.gfsh.GfshRule;
+import org.apache.geode.test.rules.gfsh.GfshScript;
 import org.apache.geode.test.junit.categories.AcceptanceTest;
 
 @Category(AcceptanceTest.class)
