+++ geode-assembly/src/test/java/org/apache/geode/management/internal/cli/commands/StatusLocatorRealGfshTest.java
@@ -18,8 +18,8 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
-import org.apache.geode.test.internal.rules.gfsh.GfshRule;
-import org.apache.geode.test.internal.rules.gfsh.GfshScript;
+import org.apache.geode.test.rules.gfsh.GfshRule;
+import org.apache.geode.test.rules.gfsh.GfshScript;
 import org.apache.geode.test.junit.categories.AcceptanceTest;
 
 @Category(AcceptanceTest.class)
