+++ geode-junit/src/main/java/org/apache/geode/test/rules/RequiresGeodeHome.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.test.internal.rules;
+package org.apache.geode.test.rules;
 
 import static org.apache.commons.lang.SystemUtils.LINE_SEPARATOR;
 import static org.assertj.core.api.Assertions.assertThat;
