+++ geode-junit/src/main/java/org/apache/geode/test/rules/gfsh/GfshRule.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.test.internal.rules.gfsh;
+package org.apache.geode.test.rules.gfsh;
 
 import static org.apache.commons.lang.SystemUtils.PATH_SEPARATOR;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -31,7 +31,7 @@ import java.util.stream.Collectors;
 import org.junit.rules.ExternalResource;
 import org.junit.rules.TemporaryFolder;
 
-import org.apache.geode.test.internal.rules.RequiresGeodeHome;
+import org.apache.geode.test.rules.RequiresGeodeHome;
 
 
 /**
