+++ geode-junit/src/main/java/org/apache/geode/test/rules/gfsh/ProcessLogger.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.test.internal.rules.gfsh;
+package org.apache.geode.test.rules.gfsh;
 
 import java.util.List;
 import java.util.Queue;
